requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
email-validator==2.1.1
alembic==1.13.1
pytest==8.1.1
httpx==0.27.0
